Use find_package_handle_standard_args and find_package_message.
psq_check_and_report_tool_version no longer takes a SUCCESS_RETURN
variable and instead errors out if FIND_REQUIRED was set. In addition,
it also takes a list of REQUIRED_VARS, which must be set in the current
scope and will be set in the parent scope on success.